requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1
pytest==7.4.3
python-dateutil==2.8.2
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23